packages/tilth-server/pyproject.toml
[project]
name = "tilth-server"
version = "0.1.0"
description = "Ingest and query gateways for tilth semantic memory"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "httpx>=0.27",
    "pydantic>=2.0",
    "qdrant-client>=1.9",
    "openai>=1.30",
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
    "pyyaml>=6.0",
    "prometheus-client>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "mypy>=1.10",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tilth_server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
